Support decimal of delay

+++ src/com/marquez/fishing/cmds/FPCmd.java
@@ -46,9 +46,9 @@ public class FPCmd implements CommandExecutor{
 					public void run() {
 						for(String cmd : FishingPlugin.startCommands) {
 							if(cmd.startsWith("delay!")) {
-								int delay = Integer.parseInt(cmd.replace("delay!", "").replace(" ", ""));
+								double delay = Double.parseDouble(cmd.replace("delay!", "").replace(" ", ""));
 								try {
-									Thread.sleep(delay*1000);
+									Thread.sleep((int)(delay*1000));
 								} catch (InterruptedException e) {
 									e.printStackTrace();
 								}
